[CBRD-25801] Problems that occur when a hidden column is generated due to ORDER BY clause in a subquery (#5810)
http://jira.cubrid.org/browse/CBRD-25801

In CBRD-24181, when optimizing the select list of a subquery, a column in ORDER BY is added. At that time, a hidden column property is added, and the query that has already been rewritten in qo_rewrite_hidden_col_as_derived() is unnessarary rewritten twice, causing an issue. Fix to remove the hidden property if the existing select list does not have it.